# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Admin Platform API
  version: "1.0.0"
  description: |
    Backend API for the Admin Platform.
    All module endpoints require **Bearer JWT** authentication and pass through a
    four-layer guard chain: `authenticateJwt -> requireRole -> requireFeatureEnabled -> requirePermission`.

    **Error envelope** (all non-2xx responses):
    ```json
    { "code": "<ErrorCode>", "message": "<human string>" }
    ```

    **Pagination envelope** (list endpoints):
    ```json
    { "items": [...], "page": 1, "limit": 50, "total": 123 }
    ```
  contact:
    name: Admin Platform
servers:
  - url: http://localhost:7002
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Two-tier JWT.  Tokens are signed with role-specific secrets
        (`JWT_SECRET_SUPER_ADMIN` / `JWT_SECRET_ADMIN`).
        Payload: `{ sub, email, role }`.

  # ─── Reusable schemas ───────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [UNAUTHORIZED, FORBIDDEN, BAD_REQUEST, NOT_FOUND, FEATURE_DISABLED, INTERNAL_ERROR]
        message:
          type: string
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ── Auth ──
...
    Conversation:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        participantUserIds: { type: array, items: { type: string } }
        status: { type: string, enum: [active, archived] }
        messageCount: { type: integer }
        lastMessageAt: { type: string, format: date-time }
        createdByUserId: { type: string }
        createdAt: { type: string, format: date-time }
    ChatMessage:
      type: object
      properties:
        _id: { type: string }
        conversationId: { type: string }
        senderUserId: { type: string }
        senderEmail: { type: string }
        content: { type: string }
        editedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }

    # ── Mailbox ──
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: MongoDB ObjectId

  # ─── Reusable responses ───────────────────────────────────────────
  responses:
...
    "403":
      description: Forbidden (role / permission / feature disabled)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/chat/conversations:
    get:
      tags: [Chat]
      summary: List conversations (participant-scoped)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Conversation" } }
    post:
      tags: [Chat]
      summary: Create conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, participantUserIds]
              properties:
                title: { type: string, minLength: 1, maxLength: 200 }
                participantUserIds: { type: array, items: { type: string }, minItems: 1, maxItems: 100 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /api/v1/chat/conversations/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Chat]
      summary: Get conversation (must be participant)
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
    patch:
      tags: [Chat]
      summary: Update conversation title
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /api/v1/chat/conversations/{id}/archive:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Chat]
      summary: Archive conversation
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /api/v1/chat/conversations/{id}/messages:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Chat]
      summary: List messages in conversation
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/ChatMessage" } }
    post:
      tags: [Chat]
      summary: Send message (active conversations only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 4000 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
  /api/v1/chat/conversations/{id}/stream:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Chat]
      summary: SSE stream for real-time messages (participant-scoped)
      description: |
        Server-Sent Events endpoint. Events: `message`, `message_edited`, `message_deleted`.
        30s heartbeat. Native EventSource cannot set headers, so pass JWT via `?token=`.
        Requires participant access or super_admin.
      parameters:
        - name: token
          in: query
          schema: { type: string }
          description: JWT token (alternative to Authorization header for SSE clients)
      responses:
        "200":
          description: SSE stream established
          content:
            text/event-stream:
              schema: { type: string }
        "403": { $ref: "#/components/responses/403" }
  /api/v1/chat/messages/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    patch:
      tags: [Chat]
      summary: Edit own message
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
    delete:
      tags: [Chat]
      summary: Delete own message
      responses:
        "204": { description: Deleted }
...